+++ test/integration/json_test.rb
@@ -3,14 +3,8 @@ require "test_helper"
 class JsonTest < IntegrationTest
   test "works with controller that inherits from ActionController::API" do
     page.driver.browser.post "/json/create-account",
-      JSON.generate({
-        "login"            => "user@example.com",
-        "password"         => "secret",
-        "password-confirm" => "secret",
-      }),
-      {
-        "CONTENT_TYPE" => "application/json",
-      }
+      { "login" => "user@example.com", "password" => "secret", "password-confirm" => "secret" }.to_json,
+      { "CONTENT_TYPE" => "application/json", "HTTP_ACCEPT" => "application/json" }
 
     assert_equal %({"success":"An email has been sent to you with a link to verify your account"}), page.html